+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Expose the Command which led to `InvalidCommandSenderException`s
  - Expose the CommandContext which led to `CommandExecutionException`s
  - Added helper methods for command flags to MutableCommandBuilder
+ - Added injection services
+ - Added a Guice injection service
+
+### Deprecated
+ - Deprecated ParameterInjectorRegistry#injectors
 
 ## [1.3.0] - 2020-12-18
 
@@ -27,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
  - String keyed command meta
+ - Deprecated ParameterInjectorRegistry#injectors
 
 ### Fixed
  - Fixed issue with task synchronization